Add an assert that there is no placeholder in a public input location (#12)
# What ❔

This PR adds an assert that there is no placeholder in a public input
location.

## Why ❔

When there is a mismatch between the setup and the witness data a
placeholder index can appear in a public input location and generate a
generic index out of bounds error. This adds an explicit check with a
descriptive error message.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Code has been formatted via `cargo fmt` and linted via `cargo
check`.