implement GpuProofConfig abstraction (#37)
# What ❔

This PR implements a GpuProofConfig abstraction to remove a dependency
on CSReferenceAssembly structure.

## Why ❔

Instantiation of the CSReferenceAssembly structure is expensive so we
need to enable the possibility to supply the pieces of configuration the
prover needs in an alternative way. This abstraction makes it possible
to generate the information either from the CSReferenceAssembly
structure or directly from a Circuit structure.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `cargo fmt` and linted via `cargo
check`.